fix: clippy cast + missing docs from fast-profile merge
- Remove redundant `as i32` cast on libc::EAGAIN in splice.rs
- Add doc comments for with_profile_and_fast, ADAPTIVE_SPLICE_MIN_BYTES,
  FastPoolPolicy, FastSplicePolicy variants, and PooledStream<S> methods

+++ crates/blackwire-common/src/splice.rs
@@ -462,7 +462,7 @@ mod linux {
                             TAG_BA_OUT => {
                                 state.ba_pending -= 1;
                                 let n = result;
-                                if n == -(libc::EAGAIN as i32) {
+                                if n == -libc::EAGAIN {
                                     self.push_out_guarded(
                                         ba_rfd,
                                         a_fd,
